Resolution problème CORS sur post-service
