+++ back/post-service/src/main/java/com/elducche/postservice/service/CommentService.java
@@ -1,0 +1,24 @@
+package com.elducche.postservice.service;
+
+import com.elducche.postservice.models.Comment;
+import com.elducche.postservice.repositories.CommentRepository;
+import org.springframework.stereotype.Service;
+import java.util.List;
+
+@Service
+public class CommentService {
+
+    private final CommentRepository commentRepository;
+
+    public CommentService(CommentRepository commentRepository) {
+        this.commentRepository = commentRepository;
+    }
+
+    public Comment createComment(Comment comment) {
+        return commentRepository.save(comment);
+    }
+
+    public List<Comment> getCommentsByPost(Long postId) {
+        return commentRepository.findByPostId(postId);
+    }
+}
