+++ back/post-service/src/main/java/com/elducche/postservice/service/SubscriptionService.java
@@ -1,0 +1,24 @@
+package com.elducche.postservice.service;
+
+import com.elducche.postservice.models.Subscription;
+import com.elducche.postservice.repositories.SubscriptionRepository;
+import org.springframework.stereotype.Service;
+import java.util.List;
+
+@Service
+public class SubscriptionService {
+
+    private final SubscriptionRepository subscriptionRepository;
+
+    public SubscriptionService(SubscriptionRepository subscriptionRepository) {
+        this.subscriptionRepository = subscriptionRepository;
+    }
+
+    public Subscription subscribe(Subscription subscription) {
+        return subscriptionRepository.save(subscription);
+    }
+
+    public List<Subscription> getSubscriptionsByUser(Long userId) {
+        return subscriptionRepository.findByUserId(userId);
+    }
+}
