--- back/post-service/src/main/java/com/elducche/postservice/services/ThemeService.java
@@ -1,24 +1,0 @@
-package com.elducche.postservice.services;
-
-import com.elducche.postservice.models.Theme;
-import com.elducche.postservice.repositories.ThemeRepository;
-import org.springframework.stereotype.Service;
-import java.util.List;
-
-@Service
-public class ThemeService {
-
-    private final ThemeRepository themeRepository;
-
-    public ThemeService(ThemeRepository themeRepository) {
-        this.themeRepository = themeRepository;
-    }
-
-    public List<Theme> getAllThemes() {
-        return themeRepository.findAll();
-    }
-
-    public Theme createTheme(Theme theme) {
-        return themeRepository.save(theme);
-    }
-}
